+++ run-on-docker.ps1
@@ -73,7 +73,7 @@ function Show-RunOnDockerHelp {
 translator Docker run - build API + web images and start the Compose stack
 
 Usage:
-  .\run-on-docker.ps1 [--ssh-string=<alias>] [--delete-volume=<no|yes>] [--network=<name>] [--api-host=<name>] [--api-port=<port>]
+  .\run-on-docker.ps1 [--ssh-string=<alias>] [--delete-volume=<no|yes>] [--network=<name>] [--api-host=<name>] [--api-port=<port>] [--help]
 
 Arguments:
   --ssh-string=<alias>        SSH config alias for remote Docker (e.g. example)
@@ -84,9 +84,11 @@ Arguments:
   --network=<name>            Docker network for the stack (default: translator-net)
   --api-host=<name>           API container hostname on that network (default: translator)
   --api-port=<port>           API port on that network for /api proxy (default: 8080)
+  --help, -h                  Show this help message and exit
 
 Examples:
   .\run-on-docker.ps1
+  .\run-on-docker.ps1 --help
   .\run-on-docker.ps1 --delete-volume=yes
   .\run-on-docker.ps1 --network=translator-net --api-port=8080
   .\run-on-docker.ps1 --ssh-string=example
@@ -161,6 +163,11 @@ function Merge-CliArguments {
         $argument = $RemainingArguments[$index]
         if ($argument -match '^--?(?<name>[\w-]+)(?:=(?<value>.*))?$') {
             $normalizedKey = ($Matches['name'] -replace '-', '_').ToLowerInvariant()
+            if ($normalizedKey -in @('help', 'h')) {
+                $merged['help'] = $true
+                $index++
+                continue
+            }
             if ($null -ne $Matches['value'] -and $Matches['value'] -ne '') {
                 $merged[$normalizedKey] = Remove-SurroundingQuotes -Value $Matches['value']
                 $index++
@@ -174,7 +181,7 @@ function Merge-CliArguments {
                 $index++
             }
         }
-        elseif ($argument -match '^(-help|-\?|/\?)$') {
+        elseif ($argument -match '^(-h|-help|--help|-\?|/\?)$') {
             $merged['help'] = $true
             $index++
         }
@@ -438,22 +445,39 @@ function Set-ComposeEnvironment {
     param(
         [string]$NetworkName,
         [string]$ApiHostName,
-        [string]$ApiPortNumber
+        [string]$ApiPortNumber,
+        [bool]$PublishHostPorts = $true
     )
 
     $env:DOCKER_NETWORK = $NetworkName
     $env:API_HOST = $ApiHostName
     $env:API_PORT = $ApiPortNumber
+
+    if ($PublishHostPorts) {
+        Remove-Item Env:API_PUBLISH_PORT -ErrorAction SilentlyContinue
+        Remove-Item Env:WEB_PUBLISH_PORT -ErrorAction SilentlyContinue
+        Remove-Item Env:POSTGRES_PUBLISH_PORT -ErrorAction SilentlyContinue
+    }
+    else {
+        $env:API_PUBLISH_PORT = ''
+        $env:WEB_PUBLISH_PORT = ''
+        $env:POSTGRES_PUBLISH_PORT = ''
+    }
 }
 
 function Get-RemoteComposeEnvironmentPrefix {
     param(
         [string]$NetworkName,
         [string]$ApiHostName,
-        [string]$ApiPortNumber
+        [string]$ApiPortNumber,
+        [bool]$PublishHostPorts = $false
     )
 
-    return "DOCKER_NETWORK='$NetworkName' API_HOST='$ApiHostName' API_PORT='$ApiPortNumber' "
+    $prefix = "DOCKER_NETWORK='$NetworkName' API_HOST='$ApiHostName' API_PORT='$ApiPortNumber' "
+    if (-not $PublishHostPorts) {
+        $prefix += "API_PUBLISH_PORT='' WEB_PUBLISH_PORT='' POSTGRES_PUBLISH_PORT='' "
+    }
+    return $prefix
 }
 
 function Get-RemoteWorkDir {
@@ -534,7 +558,7 @@ function Invoke-ComposeStack {
     if ($Target.IsLocal) {
         Push-Location $WorkingDirectory
         try {
-            Set-ComposeEnvironment -NetworkName $NetworkName -ApiHostName $ApiHostName -ApiPortNumber $ApiPortNumber
+            Set-ComposeEnvironment -NetworkName $NetworkName -ApiHostName $ApiHostName -ApiPortNumber $ApiPortNumber -PublishHostPorts:$Target.IsLocal
             Invoke-Expression $composeDown | Out-Null
             if ($LASTEXITCODE -ne 0) {
                 Write-Host 'Compose down skipped or partial (stack may not exist yet).' -ForegroundColor DarkYellow
@@ -555,11 +579,11 @@ function Invoke-ComposeStack {
         Write-Host "Compose down skipped: $($_.Exception.Message)" -ForegroundColor DarkYellow
     }
 
-    $envPrefix = Get-RemoteComposeEnvironmentPrefix -NetworkName $NetworkName -ApiHostName $ApiHostName -ApiPortNumber $ApiPortNumber
+    $envPrefix = Get-RemoteComposeEnvironmentPrefix -NetworkName $NetworkName -ApiHostName $ApiHostName -ApiPortNumber $ApiPortNumber -PublishHostPorts:$Target.IsLocal
     Invoke-RemoteShell -Target $Target -Command "${envPrefix}$composeUp" -WorkingDirectory $WorkingDirectory
 }
 
-if ($Help -or $SshString -match '^(-help|--help|-\?|/\?)$') {
+if ($Help -or $SshString -match '^(-h|-help|--help|-\?|/\?)$') {
     Show-RunOnDockerHelp
     Get-Help $PSCommandPath -Full
     exit 0
